Temp. workaround for bwave scaling in waveforms.py
Quick modification of the default behavior for body wave scaling - intended for the workshop -. 
I replaced the hardcoded body wave scaling with the hardcoded surface wave scaling, such that they have the same vertical proportions. Ryan's original values have been commented out so we can reintroduce them properly. I will work on a basic customization option tomorrow, but we will resort to this in case we can't make it before the deadline (April 13th, for the final docker image).