SPECFEM3D SAC and SGT clients support depth grid searches
I modified the SPECFEM3D clients to support grid searches in depth. So far, the GFs path pointed to the SAC files GFs directly. Now, similarly as the FK client, the GFs path has to point to a directory where there are many subdirectory each of them storing the GFs for a given depth: PATH_GF/evdp/SAC_FILES.

I modified the data directory and the tests scripts to work in agreement to the clients modifications.