Added station number in the header
Added a header field for N-Np-Ns which changed depending on which type of plot is used. I used a decorator to append additional header information to the header class, which gets written on plot if present.
They will be present by default, but the decorator make it easy to remove if needed.