Added multi-beachball plotting functionality to _plot_beachball_matplotlib
This commit integrates two different features:
Adding the option to input several moment tensors to plot on a map fashion (used in lune plot and depth plot).
Adding a "Level of detail" kind of optimization to adapt rendering speed depending on the size of the plotted beachball (single figure, or multiple beachball on a plot).