Implemented beachball plot for depth/misfit using matplotlib
The plot requires using a double axes trick, as the beachball pattern are not converted into markers or symbol, they can be stretched if the axes are not the same order of magnitude.
A secondary axis with unit extent (-1,1 / -1,1) is overlayed on the base figure, and the coordinates of the points (depths/values) are re-scaled accordingly, so that the beachballs are always perfect circles.